Add polyglotko code for train code

+++ applications/Chat/coati/models/polyglotko/polyglotko_rm.py
@@ -1,7 +1,10 @@
 from typing import Optional
+import torch
 
 import torch.nn as nn
-from transformers import GPTNeoXConfig, GPTNeoXModel
+from transformers.models.gpt_neox.configuration_gpt_neox import GPTNeoXConfig
+from transformers.models.gpt_neox.modeling_gpt_neox import GPTNeoXForCausalLM, GPTNeoXModel
+
 
 from ..base import RewardModel
 
@@ -25,7 +28,8 @@ class PolyglotkoRM(RewardModel):
         lora_train_bias: str = "none",
     ) -> None:
         if pretrained is not None:
-            model = GPTNeoXModel.from_pretrained(pretrained)
+            model = GPTNeoXModel.from_pretrained(pretrained, 
+            torch_dtype=torch.float16)
         elif config is not None:
             model = GPTNeoXModel(config)
         else:
